moving varStatement code to separate file!!
