+++ ast/ast.go
@@ -26,16 +26,6 @@ func (p *Program) TokenLiteral() string {
 	}
 }
 
-type VarStatement struct {
-	Token token.Token
-	Name *Identifier
-	Value Expression
-}
-
-func (ls *VarStatement) statementNode() {}
-
-func (ls *VarStatement) TokenLiteral() string { return ls.Token.Literal }
-
 type Identifier struct {
 	Token token.Token
 	Value string
